# contracts/openapi/nexus-skill-receiver-v1.yaml
openapi: 3.1.0
info:
  title: DeerFlow Nexus Skill Receiver API
  version: 1.0.0
  summary: Canonical service-to-service contract for Nexus-managed Skill installation in DeerFlow.
  description: >-
    This is the only canonical Nexus Skill receiver contract owned by DeerFlow. It is a
    service-to-service API for one Nexus binding; browsers must use the Nexus public API.
    Version 1 defines capability discovery, a controlled user directory, first-install
    commands for native GLOBAL and exact USER targets, durable operations, and scope-specific
    Observed state. It does not claim that the provider implementation or release-gated
    authentication, privacy, global activation, trust, compatibility, or recovery policies
    are available. Until those gates are approved, providers advertise read_only or unsupported.
  license:
    name: Apache-2.0
    identifier: Apache-2.0
servers:
  - url: /
    description: The configured DeerFlow Gateway origin; deployment endpoints are never package authority.
...
paths:
  /api/v1/nexus/skill-receiver/capabilities:
    get:
      operationId: getNexusSkillReceiverCapabilities
      summary: Read independent receiver capability and authorization dimensions
      description: >-
        Returns a snapshot relative to the authenticated machine principal. A healthy connection
        never implies write support. Every write additionally requires read_write, current freshness,
        the matching intrinsic capability, the matching granted service action, durable operations,
        exact package-digest observation, and all release gates. Unapproved service authentication
        forces read_only or unsupported.
      tags: [Nexus Skill Receiver]
      security:
        - receiverServiceBearer: []
        - receiverMutualTls: []
      x-required-service-actions:
        - receiver:capabilities:read
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Current capability snapshot for the authenticated receiver service principal.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverCapabilitySnapshot'
              examples:
                defaultDeny:
                  summary: Reachable receiver with unresolved release gates
                  value:
                    contractVersion: 1.0.0
                    transportProfile: ssh_v1
                    runtimeVersion: 0.9.0
                    connection: healthy
                    accessMode: read_only
                    freshness: current
                    authorization:
                      profile: unapproved
                      grantedActions: []
                    capabilities:
                      userDirectory: unsupported
                      globalInstall: unsupported
                      userInstall: unsupported
                      observation: unsupported
                      activation: unsupported
                      durableOperations: unsupported
                      observedPackageDigest: unsupported
                    blockedBy:
                      - CAPABILITY_READ_ONLY
                      - USER_DIRECTORY_UNSUPPORTED
                      - GLOBAL_INSTALL_UNSUPPORTED
                      - USER_INSTALL_UNSUPPORTED
                    observedAt: '2026-08-11T04:00:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/nexus/skill-receiver/users:
    get:
      operationId: listNexusSkillReceiverUsers
      summary: Search the controlled DeerFlow user directory
      description: >-
        Returns a bounded, cursor-paginated, minimal projection. The opaque cursor is bound to
        the normalized search and authorization context; the provider must not log raw search
        text or expose a complete email address. This operation is not an exact-user lookup and
        does not accept a manually supplied user ID. A USER install must revalidate the selected
        opaque ID when the command is submitted. Unauthorized callers cannot enumerate users.
      tags: [Nexus Skill Receiver]
      security:
        - receiverServiceBearer: []
        - receiverMutualTls: []
      x-required-service-actions:
        - receiver:user-directory:read
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/DirectoryQuery'
        - $ref: '#/components/parameters/Cursor'
        - $ref: '#/components/parameters/Limit'
      responses:
        '200':
          description: Stable page of minimal user projections for the current search snapshot.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverUserPage'
              examples:
                eligibleUsers:
                  value:
                    items:
                      - userId: df-user-42
                        displayName: Chen Wei
                        accountLabel: c***@example.com
                        active: true
                        installEligible: true
                    page:
                      hasMore: false
                      nextCursor: null
                    observedAt: '2026-08-11T04:01:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/nexus/skill-receiver/operations:
    post:
      operationId: createNexusSkillReceiverOperation
      summary: Submit a durable first-install operation for GLOBAL or an exact USER
      description: >-
        Nexus pre-generates receiverOperationId and sends the immutable package bytes. For USER,
        DeerFlow revalidates that the opaque directory user exists, is active, and remains eligible
        immediately before mutation. For GLOBAL, DeerFlow must use a native global Skill space and
        must not enumerate users. The Idempotency-Key, X-Request-SHA256, receiverOperationId, and
        canonical command are one durable binding. A same-key replay returns the original outcome;
        changing the hash, operation ID, or canonical command returns IDEMPOTENCY_KEY_REUSED.
        HTTP 202 only proves durable acceptance, never installation success.
      tags: [Nexus Skill Receiver]
      security:
        - receiverServiceBearer: []
        - receiverMutualTls: []
      x-required-service-actions-by-target:
        GLOBAL:
          - receiver:install:global
        USER:
          - receiver:install:user
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
        - $ref: '#/components/parameters/IdempotencyKey'
        - $ref: '#/components/parameters/RequestSha256'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/ReceiverInstallMultipartRequest'
            encoding:
              command:
                contentType: application/json
              package:
                contentType: application/zip
            examples:
              globalFirstInstall:
                summary: Native GLOBAL first install
                value:
                  command:
                    receiverOperationId: 39d78cd3-c529-4503-90f4-5077e3aebf91
                    receiverBindingId: df-primary-prod
                    skillVersionId: sv.research-assistant.1.0.0
                    runtimeSkillName: research-assistant
                    packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                    packageSizeBytes: 12345
                    packageMediaType: application/zip
                    policyRevision: trust-policy-2026-08-11
                    expectedObservedState: ABSENT
                    actorAudit:
                      principalId: 11111111-1111-4111-8111-111111111111
                      action: 'skill:install_global'
                    target:
                      scope: GLOBAL
                  package: '<binary Skill archive>'
              selectedUserFirstInstall:
                summary: USER first install after directory selection
                value:
                  command:
                    receiverOperationId: 49d78cd3-c529-4503-90f4-5077e3aebf92
                    receiverBindingId: df-primary-prod
                    skillVersionId: sv.research-assistant.1.0.0
                    runtimeSkillName: research-assistant
                    packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                    packageSizeBytes: 12345
                    packageMediaType: application/zip
                    policyRevision: trust-policy-2026-08-11
                    expectedObservedState: ABSENT
                    actorAudit:
                      principalId: 11111111-1111-4111-8111-111111111111
                      action: 'skill:install_for_user'
                    target:
                      scope: USER
                      deerFlowUserId: df-user-42
                  package: '<binary Skill archive>'
      responses:
        '202':
          description: The receiver durably bound and accepted the operation; the phase is not success.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Location:
              $ref: '#/components/headers/OperationLocation'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverOperation'
              examples:
                acceptedUserInstall:
                  value:
                    operationId: 49d78cd3-c529-4503-90f4-5077e3aebf92
                    phase: accepted
                    freshness: current
                    target:
                      scope: USER
                      deerFlowUserId: df-user-42
                    skillVersionId: sv.research-assistant.1.0.0
                    runtimeSkillName: research-assistant
                    packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                    observed: null
                    acceptedAt: '2026-08-11T04:02:00Z'
                    updatedAt: '2026-08-11T04:02:00Z'
                    completedAt: null
                    error: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '413':
          $ref: '#/components/responses/PackageTooLarge'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '429':
          $ref: '#/components/responses/RateLimited'
        '503':
          $ref: '#/components/responses/Unavailable'
        '504':
          $ref: '#/components/responses/Timeout'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/nexus/skill-receiver/operations/{operationId}:
    get:
      operationId: getNexusSkillReceiverOperation
      summary: Poll or recover a durable receiver operation
      description: >-
        This is the first recovery action after an unknown POST result. Polling never changes
        a phase because of timeout or transport failure. rejected, failed, and succeeded are
        terminal. A succeeded operation retains the exact target, version identity, digest,
        enabled state, and loaded activation evidence that proved success.
      tags: [Nexus Skill Receiver]
      security:
        - receiverServiceBearer: []
        - receiverMutualTls: []
      x-required-service-actions:
        - receiver:operations:read
      parameters:
        - $ref: '#/components/parameters/OperationId'
        - $ref: '#/components/parameters/CorrelationId'
      responses:
        '200':
          description: Current durable phase and the latest scope-specific Observed snapshot.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
            Retry-After:
              $ref: '#/components/headers/RetryAfter'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiverOperation'
              examples:
                succeededForSelectedUser:
                  value:
                    operationId: 49d78cd3-c529-4503-90f4-5077e3aebf92
                    phase: succeeded
                    freshness: current
                    target:
                      scope: USER
                      deerFlowUserId: df-user-42
                    skillVersionId: sv.research-assistant.1.0.0
                    runtimeSkillName: research-assistant
                    packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                    observed:
                      target:
                        scope: USER
                        deerFlowUserId: df-user-42
                      presence: installed
                      skillVersionId: sv.research-assistant.1.0.0
                      version: 1.0.0
                      packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                      runtimeSkillName: research-assistant
                      enabled: true
                      loadState: loaded
                      freshness: current
                      observedAt: '2026-08-11T04:04:00Z'
                    acceptedAt: '2026-08-11T04:02:00Z'
                    updatedAt: '2026-08-11T04:04:00Z'
                    completedAt: '2026-08-11T04:04:00Z'
                    error: null
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
  /api/v1/nexus/skill-receiver/observations/query:
    post:
      operationId: queryNexusSkillReceiverObservation
      summary: Read exact Observed state for one scope, target, and runtime Skill name
      description: >-
        This read-only POST carries a closed target union in JSON so GLOBAL can never contain a
        user ID and USER can never omit one. Global state is never expanded into per-user state,
        and one user's observation cannot prove another user or GLOBAL. Presence alone is not
        activation success: the exact installed identity and digest must also be enabled and loaded.
      tags: [Nexus Skill Receiver]
      security:
        - receiverServiceBearer: []
        - receiverMutualTls: []
      x-required-service-actions-by-target:
        GLOBAL:
          - receiver:observe:global
        USER:
          - receiver:observe:user
      parameters:
        - $ref: '#/components/parameters/CorrelationId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ObservationQuery'
            examples:
              globalSkill:
                value:
                  target:
                    scope: GLOBAL
                  runtimeSkillName: research-assistant
              selectedUserSkill:
                value:
                  target:
                    scope: USER
                    deerFlowUserId: df-user-42
                  runtimeSkillName: research-assistant
      responses:
        '200':
          description: Exact scope-specific Observed state, including absence.
          headers:
            X-Correlation-ID:
              $ref: '#/components/headers/CorrelationId'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ObservedSkillInstallation'
              examples:
                installedForSelectedUser:
                  value:
                    target:
                      scope: USER
                      deerFlowUserId: df-user-42
                    presence: installed
                    skillVersionId: sv.research-assistant.1.0.0
                    version: 1.0.0
                    packageDigest: 'sha256:aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa'
                    runtimeSkillName: research-assistant
                    enabled: true
                    loadState: loaded
                    freshness: current
                    observedAt: '2026-08-11T04:04:00Z'
                absentFromGlobal:
                  value:
                    target:
                      scope: GLOBAL
                    presence: absent
                    skillVersionId: null
                    version: null
                    packageDigest: null
                    runtimeSkillName: research-assistant
                    enabled: null
                    loadState: absent
                    freshness: current
                    observedAt: '2026-08-11T04:04:00Z'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '503':
          $ref: '#/components/responses/Unavailable'
        '500':
          $ref: '#/components/responses/InternalError'
components:
  securitySchemes:
    receiverServiceBearer:
      type: http
      scheme: bearer
      bearerFormat: OAuth2 access token
      description: >-
        Dedicated machine-principal credential from an approved profile. Browser cookies, DeerFlow
        user passwords, and generic internal tokens are forbidden. Publishing this scheme does not
        approve the OAuth2 client-credentials release gate.
    receiverMutualTls:
      type: mutualTLS
      description: >-
        Dedicated machine identity from an approved mTLS or combined profile. Publishing this
        scheme does not approve certificate issuance, rotation, revocation, or service scopes.
  parameters:
    CorrelationId:
      name: X-Correlation-ID
      in: header
      required: true
      description: Effective Nexus correlation ID, unchanged for the durable operation.
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:/-]+$'
...
    DirectoryQuery:
      name: query
      in: query
      required: false
      description: Bounded name or contract-approved masked-account search; never logged or echoed raw.
      schema:
        type: string
        minLength: 1
        maxLength: 100
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque cursor bound to the normalized query and authorization context.
      schema:
        type: string
        minLength: 1
        maxLength: 500
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 25
  headers:
    CorrelationId:
      description: Effective correlation ID returned unchanged or generated by the trusted caller boundary.
      required: true
      schema:
        type: string
        minLength: 1
        maxLength: 128
        pattern: '^[A-Za-z0-9._:/-]+$'
    OperationLocation:
      description: Canonical URI for the pre-generated durable receiver operation.
      required: true
      schema:
        type: string
        format: uri-reference
...
  schemas:
...
    ReceiverCapabilitySnapshot:
      type: object
      additionalProperties: false
      required:
        - contractVersion
        - transportProfile
        - runtimeVersion
        - connection
        - accessMode
        - freshness
        - authorization
        - capabilities
        - blockedBy
        - observedAt
      properties:
        contractVersion:
          $ref: '#/components/schemas/SemVer'
        transportProfile:
          $ref: '#/components/schemas/ReceiverTransportProfile'
        runtimeVersion:
          oneOf:
            - type: string
              minLength: 1
              maxLength: 100
            - type: 'null'
        connection:
          $ref: '#/components/schemas/ReceiverConnectionStatus'
        accessMode:
          $ref: '#/components/schemas/ReceiverAccessMode'
        freshness:
          $ref: '#/components/schemas/ReceiverFreshness'
        authorization:
          $ref: '#/components/schemas/ReceiverAuthorizationSnapshot'
        capabilities:
          $ref: '#/components/schemas/ReceiverCapabilities'
        blockedBy:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ReceiverErrorCode'
        observedAt:
          oneOf:
            - type: string
              format: date-time
            - type: 'null'
      not:
        anyOf:
          - properties:
              accessMode:
                const: read_write
              authorization:
                properties:
                  profile:
                    const: unapproved
                required: [profile]
            required: [accessMode, authorization]
          - properties:
              transportProfile:
                const: ssh_v1
              authorization:
                properties:
                  profile:
                    enum: [oauth2_client_credentials, mtls, combined]
                required: [profile]
            required: [transportProfile, authorization]
          - properties:
              transportProfile:
                const: http_v1
              authorization:
                properties:
                  profile:
                    const: ssh_forced_command
                required: [profile]
            required: [transportProfile, authorization]
...
    ReceiverUserPage:
      type: object
      additionalProperties: false
      required: [items, page, observedAt]
      properties:
        items:
          type: array
          maxItems: 100
          items:
            $ref: '#/components/schemas/ReceiverUser'
        page:
          $ref: '#/components/schemas/CursorPage'
        observedAt:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Dedicated receiver machine authentication is absent or invalid.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ReceiverProblem'
          examples:
            authenticationRequired:
              value:
                type: /problems/authentication-required
                title: Authentication required
                status: 401
                code: AUTHENTICATION_REQUIRED
                detail: A dedicated receiver service identity is required.
                instance: /api/v1/nexus/skill-receiver/problems/01J5AUTHN
                correlationId: receiver-authn-01
                timestamp: '2026-08-11T04:05:00Z'
                retryable: false
    Forbidden:
      description: The authenticated machine principal lacks a required receiver action.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ReceiverProblem'
          examples:
            serviceActionMissing:
              value:
                type: /problems/forbidden
                title: Receiver action forbidden
                status: 403
                code: FORBIDDEN
                detail: The receiver service principal lacks the required action.
                instance: /api/v1/nexus/skill-receiver/problems/01J5AUTHZ
                correlationId: receiver-authz-01
                timestamp: '2026-08-11T04:05:00Z'
                retryable: false
...
    Conflict:
      description: Capability, target eligibility, first-install state, or idempotency conflict.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ReceiverProblem'
          examples:
            idempotencyKeyReused:
              value:
                type: /problems/idempotency-key-reused
                title: Idempotency key already bound
                status: 409
                code: IDEMPOTENCY_KEY_REUSED
                detail: The idempotency key is bound to another operation or canonical request.
                instance: /api/v1/nexus/skill-receiver/problems/01J5IDEMPOTENCY
                correlationId: skill-install-20260811-02
                timestamp: '2026-08-11T04:05:00Z'
                retryable: false
            targetUserNotEligible:
              value:
                type: /problems/target-user-not-eligible
                title: Target user is not eligible
                status: 409
                code: TARGET_USER_NOT_ELIGIBLE
                detail: The selected user cannot receive Skill installations.
                instance: /api/v1/nexus/skill-receiver/problems/01J5TARGET
                correlationId: skill-install-20260811-03
                timestamp: '2026-08-11T04:05:00Z'
                retryable: false
...
    Unavailable:
      description: Receiver or requested provider capability is temporarily unavailable or not ready.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
        Retry-After:
          $ref: '#/components/headers/RetryAfter'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ReceiverProblem'
...
    InternalError:
      description: Unclassified receiver failure with restricted diagnostics.
      headers:
        X-Correlation-ID:
          $ref: '#/components/headers/CorrelationId'
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/ReceiverProblem'